Create jupyter namespace for Jupyter-specific epics
